feat(auth): password change
Changed the 'password changed' implementation to the authentication.service file
