+++ src/aws/aws.service.ts
@@ -60,4 +60,5 @@ export class AwsService {
         throw err;
       }
     }
+
 }
